Initial Analysis
Made two analyses:
1. pricePeaks - This was initially made to test the three highest peaks for short interest in the market, note their dates, then track the share price for the nexxt 6 months following the peak. this could be used to advise a trading strategy.
however, i could not find a free provider that gave short-interest data. I tried Alpha Vantage, IEX (now no longer exists) and Finnhub.
Since this data was unavailable, i have made a checkpoint and used it on share price data instead, with the intention to switch to short-interest when a provider is discovered.

2. A insider trading script was also made to track the net insider trading for requested stocks over the last year, returning graphs for the user.
Since Alpha Vantage has a limit of 25 transactions a day, i will have to return to this tomorrow as i am not receiving any data now.

ToDo:
 - Find a short-interest provider and integrate into pricePeaks
 - Test insiderTransactions another day when the 25 request API limit is refreshed.